requirements.txt
torch>=1.8.0
torchvision>=0.9.0
datasets>=2.0.0
tqdm>=4.62.0
matplotlib>=3.4.0
numpy>=1.20.0
pillow>=8.2.0
scikit-learn>=1.0.0